Move bucket-mismatched test fixtures to their correct component
