+++ tests/Console/Scheduling/ScheduleTest.php
@@ -9,7 +9,7 @@ use Illuminate\Console\Scheduling\Schedule;
 use Illuminate\Console\Scheduling\SchedulingMutex;
 use Illuminate\Container\Container;
 use Illuminate\Contracts\Queue\ShouldQueue;
-use Illuminate\Tests\Console\Fixtures\JobToTestWithSchedule;
+use Illuminate\Tests\Queue\Fixtures\JobToTestWithSchedule;
 use Mockery;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\DataProvider;
