+++ tests/Integration/Database/EloquentTransactionWithAfterCommitTests.php
@@ -2,13 +2,9 @@
 
 namespace Illuminate\Tests\Integration\Database;
 
-use Illuminate\Bus\Queueable;
-use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Foundation\Auth\User;
-use Illuminate\Foundation\Bus\Dispatchable;
-use Illuminate\Queue\InteractsWithQueue;
-use Illuminate\Support\Carbon;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Tests\Queue\Fixtures\EloquentTransactionWithAfterCommitTestsJob;
 use Orchestra\Testbench\Concerns\WithLaravelMigrations;
 use Orchestra\Testbench\Factories\UserFactory;
 
@@ -210,25 +206,6 @@ class EloquentTransactionWithAfterCommitTestsUserObserverUsingDispatchSync exten
         dispatch_sync(new EloquentTransactionWithAfterCommitTestsJob($user->email));
 
         parent::created($user);
-    }
-}
-
-class EloquentTransactionWithAfterCommitTestsJob implements ShouldQueue
-{
-    use Dispatchable, InteractsWithQueue, Queueable;
-
-    public function __construct(public string $email)
-    {
-        // ...
-    }
-
-    public function handle(): void
-    {
-        DB::transaction(function () {
-            DB::table('password_reset_tokens')->insert([
-                ['email' => $this->email, 'token' => sha1($this->email), 'created_at' => Carbon::now()],
-            ]);
-        });
     }
 }
