+++ tests/Integration/Http/Fixtures/ModelInspectorTestModelResource.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Illuminate\Tests\Integration\Http\Fixtures;
+
+use Illuminate\Http\Resources\Json\JsonResource;
+
+class ModelInspectorTestModelResource extends JsonResource
+{
+}
