+++ tests/Integration/Queue/DeleteNotificationWhenMissingModelTest.php
@@ -2,16 +2,12 @@
 
 namespace Illuminate\Tests\Integration\Queue;
 
-use Illuminate\Bus\Queueable;
-use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Notifications\Notifiable;
-use Illuminate\Notifications\Notification;
-use Illuminate\Queue\Attributes\DeleteWhenMissingModels;
-use Illuminate\Queue\SerializesModels;
 use Illuminate\Support\Facades\Notification as NotificationFacade;
 use Illuminate\Support\Facades\Schema;
+use Illuminate\Tests\Notifications\Fixtures\DeleteWhenMissingNotification;
 use Orchestra\Testbench\Attributes\WithMigration;
 
 #[WithMigration]
@@ -69,28 +65,4 @@ class DeleteNotificationTestModel extends Model
     public $timestamps = false;
 
     protected $guarded = [];
-}
-
-#[DeleteWhenMissingModels]
-class DeleteWhenMissingNotification extends Notification implements ShouldQueue
-{
-    use Queueable, SerializesModels;
-
-    public static bool $sent = false;
-
-    public function __construct(public DeleteNotificationTestModel $model)
-    {
-    }
-
-    public function via($notifiable): array
-    {
-        return ['mail'];
-    }
-
-    public function toMail($notifiable)
-    {
-        static::$sent = true;
-
-        return new \Illuminate\Notifications\Messages\MailMessage;
-    }
 }
