+++ tests/Integration/Support/DeferredCallbackTest.php
@@ -2,11 +2,9 @@
 
 namespace Illuminate\Tests\Integration\Support;
 
-use Illuminate\Bus\Queueable;
-use Illuminate\Contracts\Queue\ShouldQueue;
-use Illuminate\Foundation\Bus\Dispatchable;
 use Illuminate\Foundation\Http\Middleware\InvokeDeferredCallbacks;
 use Illuminate\Support\Facades\Route;
+use Illuminate\Tests\Queue\Fixtures\TestSyncJob;
 use Orchestra\Testbench\Attributes\WithConfig;
 use Orchestra\Testbench\TestCase;
 
@@ -28,15 +26,5 @@ class DeferredCallbackTest extends TestCase
         $this->get('/test');
 
         $this->assertTrue($executed);
-    }
-}
-
-class TestSyncJob implements ShouldQueue
-{
-    use Dispatchable, Queueable;
-
-    public function handle(): void
-    {
-        //
     }
 }
