+++ tests/Queue/Fixtures/EloquentTransactionWithAfterCommitTestsJob.php
@@ -1,0 +1,29 @@
+<?php
+
+namespace Illuminate\Tests\Queue\Fixtures;
+
+use Illuminate\Bus\Queueable;
+use Illuminate\Contracts\Queue\ShouldQueue;
+use Illuminate\Foundation\Bus\Dispatchable;
+use Illuminate\Queue\InteractsWithQueue;
+use Illuminate\Support\Carbon;
+use Illuminate\Support\Facades\DB;
+
+class EloquentTransactionWithAfterCommitTestsJob implements ShouldQueue
+{
+    use Dispatchable, InteractsWithQueue, Queueable;
+
+    public function __construct(public string $email)
+    {
+        // ...
+    }
+
+    public function handle(): void
+    {
+        DB::transaction(function () {
+            DB::table('password_reset_tokens')->insert([
+                ['email' => $this->email, 'token' => sha1($this->email), 'created_at' => Carbon::now()],
+            ]);
+        });
+    }
+}
