+++ tests/Queue/Fixtures/JobToTestWithSchedule.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Illuminate\Tests\Console\Fixtures;
+namespace Illuminate\Tests\Queue\Fixtures;
 
 use Illuminate\Contracts\Queue\ShouldQueue;
